move Exception and Registry to contract package
