+++ src/main/scala/com/dzegel/DynamockServer/contract/Expectation.scala
@@ -1,3 +1,3 @@
-package com.dzegel.DynamockServer.registry
+package com.dzegel.DynamockServer.contract
 
 case class Expectation(path: String, method: String, content: String)
