+++ src/main/scala/com/dzegel/DynamockServer/contract/Response.scala
@@ -1,0 +1,3 @@
+package com.dzegel.DynamockServer.contract
+
+case class Response()
