+++ src/main/scala/com/dzegel/DynamockServer/contract/SetupExpectationPostRequest.scala
@@ -1,5 +1,3 @@
 package com.dzegel.DynamockServer.contract
 
-import com.dzegel.DynamockServer.registry.{Expectation, Response}
-
 case class SetupExpectationPostRequest(expectation: Expectation, response: Response)
