--- src/main/scala/com/dzegel/DynamockServer/registry/Response.scala
@@ -1,3 +1,0 @@
-package com.dzegel.DynamockServer.registry
-
-case class Response()
