+++ src/test/scala/com/dzegel/DynamockServer/service/SetupServiceTests.scala
@@ -1,7 +1,7 @@
 package com.dzegel.DynamockServer.service
 
-import com.dzegel.DynamockServer.registry.{Expectation, Response, SetupRegistry}
-import com.dzegel.DynamockServer.contract.SetupExpectationPostRequest
+import com.dzegel.DynamockServer.registry.SetupRegistry
+import com.dzegel.DynamockServer.contract.{Expectation, Response, SetupExpectationPostRequest}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.{FunSuite, Matchers}
